+++ adept/agents/impala/actor_critic_vtrace.py
@@ -20,7 +20,7 @@ from collections import OrderedDict
 import torch
 from torch.nn import functional as F
 
-from adept.environments import Engines
+from adept.environments.registry import Engines
 from adept.expcaches.rollout import RolloutCache
 from adept.utils.util import listd_to_dlist, dlist_to_listd
 from adept.networks._base import ModularNetwork
@@ -45,7 +45,7 @@ class ActorCriticVtrace(Agent):
         self._action_keys = list(sorted(action_space.entries_by_name.keys()))
         self._func_id_to_headnames = None
         if self.engine == Engines.SC2:
-            from adept.environments.sc2 import SC2ActionLookup
+            from adept.environments.deepmind_sc2 import SC2ActionLookup
             self._func_id_to_headnames = SC2ActionLookup()
 
     @classmethod
