+++ adept/utils/script_helpers.py
@@ -14,19 +14,19 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-import os
 from argparse import ArgumentParser  # for type hinting
 
 from adept.agents import AGENTS
-from adept.environments import SubProcEnv, SC2_ENVS, Engines, DummyVecEnv
-from adept.environments import reward_normalizer_by_env_id
-from adept.environments.atari import make_atari_env
+from adept.environments import ParallelEnvManager, DebugEnvManager, SC2_ENVS
+from adept.environments._base import reward_normalizer_by_env_id
+from adept.environments.registry import Engines
+from adept.environments.openai_gym import make_atari_env
 from adept.networks import VISION_NETWORKS, DISCRETE_NETWORKS, NETWORK_BODIES
 from adept.networks._base import NetworkTrunk, ModularNetwork, NetworkHead
 from adept.utils.util import parse_bool
 
 try:
-    from adept.environments.sc2 import make_sc2_env, SC2AgentOverrides
+    from adept.environments.deepmind_sc2 import make_sc2_env, SC2AgentOverrides
 except ImportError:
     print('SC2 Environment not detected')
 
@@ -41,15 +41,15 @@ def make_env(args, seed, subprocess=True, render=False):
 
 def sc2_from_args(args, seed, subprocess=True, render=False):
     if subprocess:
-        return SubProcEnv([make_sc2_env(args.env_id, seed + i) for i in range(args.nb_env)], Engines.SC2)
+        return ParallelEnvManager([make_sc2_env(args.env_id, seed + i) for i in range(args.nb_env)], Engines.SC2)
     else:
-        return DummyVecEnv([make_sc2_env(args.env_id, seed + i, render=render) for i in range(args.nb_env)], Engines.SC2)
+        return DebugEnvManager([make_sc2_env(args.env_id, seed + i, render=render) for i in range(args.nb_env)], Engines.SC2)
 
 
 def atari_from_args(args, seed, subprocess=True):
     do_frame_stack = 'Linear' in args.network_body
 
-    env_wrapper_class = SubProcEnv if subprocess else DummyVecEnv
+    env_wrapper_class = ParallelEnvManager if subprocess else DebugEnvManager
     envs = env_wrapper_class(
         [
             make_atari_env(
